Report data transfer rates
Split out utils into separate modules
